♻️Refactor:type공유 및 오류디버깅
