+++ src/components/interest-heart/logic/putInterest.ts
@@ -1,11 +1,6 @@
 import axiosInstance from '@/api/apis';
+import { interestType } from '@/shared/types/interest/interestType';
 import { resType } from '@/shared/types/resType';
-
-interface interestType {
-  id: number;
-  volunteer_id: string;
-  center_id: string;
-}
 
 export const postInterest = async (center_id: string) => {
   try {
