--- src/components/search-bar/nonfilter-search/logic/testFunc.ts
@@ -1,5 +1,0 @@
-const testFunc = () => {
-    console.log('검색기능');
-}
-
-export default testFunc
