+++ src/features/aidreq-detail-center-profile/_components/picture-container/index.tsx
@@ -6,7 +6,7 @@ interface PictureContainerProps {
 }
 
 const PictureContainer: React.FC<PictureContainerProps> = ({ centerProfile }) => {
-  return <ProfileImg imgurl={centerProfile.img_url || ''}></ProfileImg>;
+  return <ProfileImg imgurl={centerProfile.img_url || '/assets/imgs/no-img-center.svg'}></ProfileImg>;
 };
 
 export default PictureContainer;
