+++ src/features/login-feature/_components/person-login/index.tsx
@@ -1,22 +1,8 @@
-// import { useLoginStore } from '@/store/stores/login/loginStore';
-// import { handleLoginPerson } from '@/store/queries/login/useLoginVolunteer';
 import { PersonLoginCss } from './indexCss';
 
 const PersonLogin = () => {
-  // const setLoginInfo = useLoginStore((state) => state.setLoginInfo);
-
   const onClickNaver = async () => {
     window.location.href = `${import.meta.env.VITE_APP_BASE_URL}/oauth2/authorization/naver`;
-    // setLoginInfo('persontestid', 'person');
-
-    //테스트용 함수
-    // try {
-    //   await handleLoginPerson();
-    //   setLoginInfo('9a524bd3-2542-47f1-987f-8141704a6590', 'person');
-    // } catch (error) {
-    //   console.error('Login failed:', error);
-    // }
-    //테스트용 함수
   };
 
   return (
