+++ src/shared/types/aidrq-detail/PresentResponse.ts
@@ -1,0 +1,4 @@
+export interface PresentResponse {
+  status: string;
+  attended: boolean;
+}
