+++ src/shared/types/interest/interestType.ts
@@ -1,0 +1,5 @@
+export interface interestType {
+  id: number;
+  volunteer_id: string;
+  center_id: string;
+}
