+++ src/shared/types/location/locationType.ts
@@ -1,0 +1,6 @@
+export interface LocationData {
+  region: string;
+  address: string;
+  latitude: number;
+  longitude: number;
+}
