+++ src/store/queries/aidreq-detail-volunteer-query/useCurrentRecruitment.ts
@@ -2,5 +2,5 @@ import axiosInstance from '@/api/apis';
 
 export const fetchCurrentRecruitment = async (id: number) => {
   const response = await axiosInstance.get(`/api/volunteer-apply/recruit-board/${id}/summary`);
-  return response.data.data;
+  return response.data;
 };
