+++ src/store/queries/aidreq-detail-volunteer-query/useFetchAidRqDetail.ts
@@ -1,22 +1,9 @@
-import { Dispatch, SetStateAction } from 'react';
 import axios from 'axios';
 
-import { AidRqDetailType } from '@/shared/types/aidrq-detail/aidrqDetailType';
-
-interface ApiResponse {
-  code: number;
-  message: string;
-  data: AidRqDetailType;
-}
-
-export const fetchAidRqDetail = async (
-  setData: Dispatch<SetStateAction<ApiResponse | null>>,
-  id: string | undefined
-) => {
+export const fetchAidRqDetail = async (id: string | undefined) => {
   try {
     const response = await axios.get(`${import.meta.env.VITE_APP_BASE_URL}/api/recruit-board/${id}`);
-    setData(response.data);
-    return response;
+    return response.data.data;
   } catch (error) {
     console.error('Error:', error);
   }
